Move HistoricalPegoutSelectionsConstants to their own package
