+++ rskj-core/src/main/java/co/rsk/peg/constants/BridgeMainNetConstants.java
@@ -5,6 +5,7 @@ import co.rsk.bitcoinj.core.NetworkParameters;
 import co.rsk.peg.federation.constants.FederationMainNetConstants;
 import co.rsk.peg.feeperkb.constants.FeePerKbMainNetConstants;
 import co.rsk.peg.lockingcap.constants.LockingCapMainNetConstants;
+import co.rsk.peg.pegout.HistoricalPegoutSelectionsMainNetConstants;
 import co.rsk.peg.union.constants.UnionBridgeMainNetConstants;
 import co.rsk.peg.whitelist.constants.WhitelistMainNetConstants;
