+++ rskj-core/src/main/java/co/rsk/peg/constants/BridgeRegTestConstants.java
@@ -24,6 +24,7 @@ import co.rsk.bitcoinj.core.NetworkParameters;
 import co.rsk.peg.federation.constants.FederationRegTestConstants;
 import co.rsk.peg.feeperkb.constants.FeePerKbRegTestConstants;
 import co.rsk.peg.lockingcap.constants.LockingCapRegTestConstants;
+import co.rsk.peg.pegout.HistoricalPegoutSelectionsRegTestConstants;
 import co.rsk.peg.union.constants.UnionBridgeRegTestConstants;
 import co.rsk.peg.whitelist.constants.WhitelistRegTestConstants;
 import java.util.List;
